+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 - Support for `mount_type_details`, `mount_type`, `nfs_server` and `nfs_server_export` attributes in `oci_database_backup_destination` resource
 - Support resource discovery for `ons` resources
 - Support resource discovery for `analytics` resources
+- Support resource discovery for `dns` resources
+- Support datasource for `oci_dns_rrset`
+
 ### Fixed
 - Fix issue where discovering object storage buckets without lifecycle policies, results in an error